requirements.txt
aiohttp==3.8.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==2.1.1
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
frozenlist==1.8.0
github==1.2.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
